Gym/wandb/offline-run-20220708_182135-3mv08148/files/requirements.txt
absl-py==0.13.0
alabaster==0.7.12
astor==0.8.1
astunparse==1.6.3
atari-py==0.2.6
babel==2.10.3
biwrap==0.1.6
cached-property==1.5.2
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.0.4
click==8.0.3
cloudpickle==2.0.0
configparser==5.1.0
cycler==0.10.0
cython==0.29.24
dataclasses-json==0.5.5
datetime==4.3
decorator==4.4.2
dm-tree==0.1.6
docker-pycreds==0.4.0
docutils==0.17.1
ecos==2.0.10
f110-gym==0.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gitdb==4.0.9
gitpython==3.1.24
gmpy2==2.0.8
google-auth-oauthlib==0.4.6
google-auth==1.35.0
google-pasta==0.2.0
grpcio==1.34.1
gym==0.21.0
h5py==3.1.0
idna==3.2
imageio-ffmpeg==0.4.5
imageio==2.9.0
imagesize==1.4.1
importlib-metadata==4.8.1
jinja2==3.1.2
joblib==1.0.1
jsonlines==2.0.0
keras-applications==1.0.8
keras-nightly==2.5.0.dev2021032900
keras-preprocessing==1.1.2
kiwisolver==1.3.2
llvmlite==0.38.1
markdown==3.3.4
markupsafe==2.1.1
marshmallow-enum==1.5.1
marshmallow==3.13.0
matplotlib==3.4.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpi4py==3.1.1
mpmath==1.2.1
mypy-extensions==0.4.3
nptyping==1.4.3
numba==0.55.2
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
osqp==0.6.2.post0
packaging==21.3
pandas==1.3.2
pathtools==0.1.2
pickle5==0.0.11
pillow==8.4.0
pip==21.0.1
proglog==0.1.10
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pybullet==3.1.8
pyglet==1.5.21
pygments==2.12.0
pyopengl==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pyyaml==6.0
pyzmq==22.3.0
qdldl==0.1.5.post0
qpsolvers==1.6.1
quadprog==0.1.8
racecar-gym==0.1
requests-oauthlib==1.3.0
requests==2.26.0
rsa==4.7.2
scipy==1.7.1
scs==3.2.0
sentry-sdk==1.4.3
setuptools==52.0.0.post20210125
shapely==1.7.1
shortuuid==1.0.7
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
splinify-gnone==0.0.1
stable-baselines3==1.2.0
stable-baselines==2.10.2
subprocess32==3.5.4
sympy==1.8
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboard==2.6.0
tensorflow-estimator==2.5.0
tensorflow-plot==0.3.2
tensorflow-probability==0.11.1
tensorflow==2.5.0
termcolor==1.1.0
torch==1.9.0
tqdm==4.64.0
typing-extensions==4.0.1
typing-inspect==0.7.1
typish==1.9.3
urllib3==1.26.6
wandb==0.12.6
websocket-client==1.3.3
werkzeug==2.0.1
wheel==0.37.0
wrapt==1.12.1
xvfbwrapper==0.2.9
yamldataclassconfig==1.5.0
yaspin==2.1.0
zipp==3.5.0
zope.interface==5.4.0